format and put places back in its place
