+++ packages/chrome-extension/src/app/components/common/RepoInfo.ts
@@ -1,0 +1,9 @@
+import { GitHubPageType } from "../../github/GitHubPageType";
+import { PrInfo } from "../pr/IsolatedPrEditor";
+import * as dependencies__ from "../../dependencies";
+
+export class RepoInfo {
+    public owner: string;
+    public repo: string;
+    public gitref: string;
+}
